Continued Db testing

+++ src/main/java/com/cerverae/calidad/database/DAO.java
@@ -5,6 +5,6 @@ public interface DAO {
 	public boolean deleteAlumno(Alumno a);
 	public int getAllAlumnosCount();
 	public Alumno getAlumno(int id);
-	public boolean  updateAlumnoCalificacion(Alumno a, double calificacion);
+	public boolean  updateAlumnoCalificacion(Alumno a, int calificacion);
 	
 }
